Move formatTime to utils
